+++ packages/commitlint-config/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## commitlint-config-gitmoji [2.2.4](https://github.com/arvinxx/gitmoji-commit-workflow/compare/commitlint-config-gitmoji@2.2.3...commitlint-config-gitmoji@2.2.4) (2021-05-02)
+## commitlint-config-gitmoji [2.2.5-beta.3](https://github.com/arvinxx/gitmoji-commit-workflow/compare/commitlint-config-gitmoji@2.2.5-beta.2...commitlint-config-gitmoji@2.2.5-beta.3) (2021-05-04)
 
 
 
@@ -8,7 +8,36 @@
 
 ### Dependencies
 
-* **commitlint-plugin-gitmoji:** upgraded to 2.1.1
+* **commitlint-plugin-gitmoji:** upgraded to 2.2.0-beta.3
+
+## commitlint-config-gitmoji [2.2.5-beta.2](https://github.com/arvinxx/gitmoji-commit-workflow/compare/commitlint-config-gitmoji@2.2.5-beta.1...commitlint-config-gitmoji@2.2.5-beta.2) (2021-05-02)
+
+
+### 🐛 Bug Fixes
+
+* fix dependency error ([def947f](https://github.com/arvinxx/gitmoji-commit-workflow/commit/def947f))
+
+## commitlint-config-gitmoji [2.2.5-beta.1](https://github.com/arvinxx/gitmoji-commit-workflow/compare/commitlint-config-gitmoji@2.2.4...commitlint-config-gitmoji@2.2.5-beta.1) (2021-05-02)
+
+
+
+
+
+### Dependencies
+
+* **commitlint-plugin-gitmoji:** upgraded to 2.2.0-beta.2
+
+## commitlint-config-gitmoji [2.2.4-beta.1](https://github.com/arvinxx/gitmoji-commit-workflow/compare/commitlint-config-gitmoji@2.2.3...commitlint-config-gitmoji@2.2.4-beta.1) (2021-05-02)
+
+### Dependencies
+
+- **commitlint-plugin-gitmoji:** upgraded to 2.2.0-beta.1
+
+## commitlint-config-gitmoji [2.2.4](https://github.com/arvinxx/gitmoji-commit-workflow/compare/commitlint-config-gitmoji@2.2.3...commitlint-config-gitmoji@2.2.4) (2021-05-02)
+
+### Dependencies
+
+- **commitlint-plugin-gitmoji:** upgraded to 2.1.1
 
 ## commitlint-config-gitmoji [2.2.3](https://github.com/arvinxx/gitmoji-commit-workflow/compare/commitlint-config-gitmoji@2.2.2...commitlint-config-gitmoji@2.2.3) (2021-03-06)