+++ packages/commitlint-plugin/CHANGELOG.md
@@ -1,11 +1,30 @@
 # Changelog
 
-## commitlint-plugin-gitmoji [2.1.1](https://github.com/arvinxx/commitlint-config-gitmoji/compare/commitlint-plugin-gitmoji@2.1.0...commitlint-plugin-gitmoji@2.1.1) (2021-05-02)
+# commitlint-plugin-gitmoji [2.2.0-beta.3](https://github.com/arvinxx/commitlint-config-gitmoji/compare/commitlint-plugin-gitmoji@2.2.0-beta.2...commitlint-plugin-gitmoji@2.2.0-beta.3) (2021-05-04)
 
+
+### ✨ Features
+
+* support local gitmoji path ([6b75a95](https://github.com/arvinxx/commitlint-config-gitmoji/commit/6b75a95))
+
+# commitlint-plugin-gitmoji [2.2.0-beta.2](https://github.com/arvinxx/commitlint-config-gitmoji/compare/commitlint-plugin-gitmoji@2.2.0-beta.1...commitlint-plugin-gitmoji@2.2.0-beta.2) (2021-05-02)
+
+
+### ✨ Features
+
+* support GITMOJI_PATH env ([8a81839](https://github.com/arvinxx/commitlint-config-gitmoji/commit/8a81839))
+
+# commitlint-plugin-gitmoji [2.2.0-beta.1](https://github.com/arvinxx/commitlint-config-gitmoji/compare/commitlint-plugin-gitmoji@2.1.0...commitlint-plugin-gitmoji@2.2.0-beta.1) (2021-05-02)
+
+### ✨ Features
+
+- support GITMOJI_PATH env ([2d05da1](https://github.com/arvinxx/commitlint-config-gitmoji/commit/2d05da1)), closes [#40](https://github.com/arvinxx/commitlint-config-gitmoji/issues/40)
+
+## commitlint-plugin-gitmoji [2.1.1](https://github.com/arvinxx/commitlint-config-gitmoji/compare/commitlint-plugin-gitmoji@2.1.0...commitlint-plugin-gitmoji@2.1.1) (2021-05-02)
 
 ### 🐛 Bug Fixes
 
-* fix error message reference url ([351f081](https://github.com/arvinxx/commitlint-config-gitmoji/commit/351f081))
+- fix error message reference url ([351f081](https://github.com/arvinxx/commitlint-config-gitmoji/commit/351f081))
 
 # commitlint-plugin-gitmoji [2.1.0](https://github.com/arvinxx/commitlint-config-gitmoji/compare/commitlint-plugin-gitmoji@2.0.6...commitlint-plugin-gitmoji@2.1.0) (2021-03-06)